refactor reducers
