+++ app/util/NodeUIState.ts
@@ -28,7 +28,7 @@ export function getNodeUIState(id: string, state: State): NodeUIState {
   return {
     dragged: Boolean(state.display.draggedNode === id),
     edited: Boolean(state.display.floatingEditor.type === 'node' && state.display.floatingEditor.id  === id),
-    selected: Boolean(state.display.selection['node']?.includes(id)),
+    selected: state.display.selection.node.includes(id),
     highlighted: Boolean(state.display.modes.story && state.annotations.list[state.annotations.currentIndex]?.nodeIds?.includes(id)),
     appearance: calculateAppearance(id, state)
   }
